+++ packages/skeleton-svelte/test/components/rating-group/rating-group.svelte
@@ -5,13 +5,9 @@
 <RatingGroup count={3} data-testid="root">
 	<RatingGroup.Label data-testid="label">Label</RatingGroup.Label>
 	<RatingGroup.Control data-testid="control">
-		<RatingGroup.Context>
-			{#snippet children(ctx)}
-				{#each ctx.api.items as index (index)}
-					<RatingGroup.Item {index} data-testid="item-{index}">Item</RatingGroup.Item>
-				{/each}
-			{/snippet}
-		</RatingGroup.Context>
+		{#each Array.from({ length: 5 }), index}
+			<RatingGroup.Item {index} data-testid="item-{index}" />
+		{/each}
 	</RatingGroup.Control>
 	<RatingGroup.HiddenInput data-testid="hidden-input" />
 </RatingGroup>