+++ sites/skeleton.dev/src/examples/components/rating-group/react/half.tsx
@@ -1,31 +1,14 @@
 'use client';
 
 import { RatingGroup } from '@skeletonlabs/skeleton-react';
-import { StarIcon, StarHalfIcon } from 'lucide-react';
 
-export default function Page() {
+export default function () {
 	return (
 		<RatingGroup count={5} allowHalf={true}>
 			<RatingGroup.Control>
-				<RatingGroup.Context>
-					{(ctx) =>
-						ctx.api.items.map((index) => (
-							<RatingGroup.Item key={index} index={index}>
-								<RatingGroup.ItemContext>
-									{(ctx) => {
-										if (ctx.itemState.half) {
-											return <StarHalfIcon fill="currentColor" />;
-										}
-										if (ctx.itemState.highlighted) {
-											return <StarIcon fill="currentColor" />;
-										}
-										return <StarIcon />;
-									}}
-								</RatingGroup.ItemContext>
-							</RatingGroup.Item>
-						))
-					}
-				</RatingGroup.Context>
+				{Array.from({ length: 5 }).map((_, index) => (
+					<RatingGroup.Item key={index} index={index + 1} />
+				))}
 			</RatingGroup.Control>
 			<RatingGroup.HiddenInput />
 		</RatingGroup>